requirements_melhorias.txt
# Requirements para Sistema Multi-Agente Mangaba.AI
# Inclui dependências para todas as 10 melhorias propostas

# Core Flask e extensões
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.5.3
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
Flask-Limiter==3.5.0

# Processamento assíncrono
celery==5.3.4
redis==5.0.1
eventlet==0.33.3

# Banco de dados
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
alembic==1.12.1

# Cache e performance
pymemcache==4.0.0

# Autenticação e segurança
bcrypt==4.1.2
PyJWT==2.8.0
cryptography==41.0.7

# APIs e HTTP
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# Machine Learning e NLP
scikit-learn==1.3.2
transformers==4.36.2
torch==2.1.1
numpy==1.24.4
pandas==2.1.4

# Processamento de dados
openpyxl==3.1.2
PyPDF2==3.0.1
python-docx==1.1.0

# Monitoramento e métricas
prometheus-client==0.19.0
psutil==5.9.6

# Logging estruturado
structlog==23.2.0
loguru==0.7.2

# Validação e serialização
marshmallow==3.20.1
Flask-Marshmallow==0.15.0
pydantic==2.5.2

# Utilitários
python-dotenv==1.0.0
click==8.1.7
Pillow==10.1.0
markdown==3.5.1

# Testes
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Desenvolvimento
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1

# Produção
gunicorn==21.2.0
gevent==23.9.1

# Integração com APIs externas
google-api-python-client==2.108.0
salesforce-api==0.1.47
hubspot-api-client==7.5.0

# Processamento de arquivos
chardet==5.2.0
python-magic==0.4.27

# Criptografia e hashing
hashlib-compat==1.0.1

# Timezone e data
pytz==2023.3
python-dateutil==2.8.2

# WebSocket e tempo real
python-socketio==5.10.0

# Rate limiting
Flask-Limiter==3.5.0

# Swagger/OpenAPI
flask-restx==1.3.0
flasgger==0.9.7.1

# Configuração
python-decouple==3.8

# Backup e storage
boto3==1.34.0  # AWS S3

# Análise de sentimento
textblob==0.17.1
vader-sentiment==3.3.2

# Geração de PDFs
reportlab==4.0.7
weasyprint==60.2

# Processamento de imagens
opencv-python==4.8.1.78

# Compressão
gzip==1.0

# Serialização avançada
orjson==3.9.10

# Validação de email
email-validator==2.1.0

# Internacionalização
Flask-Babel==4.0.0

# Scheduler
APScheduler==3.10.4

# Health checks
Flask-HealthCheck==1.0.0